splitup ephemeral_custom_processing
